- Fixed @unit kill requiring 'p' access if you were already 'az'.
git-svn-id: https://ascentemubackup.svn.sourceforge.net/svnroot/ascentemubackup/trunk@4281 8ad44209-8f18-4ceb-bdfa-99d0be514da2

+++ src/ascent-world/Level2.cpp
@@ -349,7 +349,7 @@ bool ChatHandler::HandleKillCommand(const char *args, WorldSession *m_session)
 	{
 	case TYPEID_PLAYER:
 		{
-			if(strchr(m_session->GetPermissions(),'p')==NULL)
+			if(strchr(m_session->GetPermissions(),'p')==NULL && !m_session->HasGMPermissions())
 			{
 				RedSystemMessage(m_session, "You do not have permission to kill a player.");
 				return true;
